requirements.txt
asgiref==3.7.2
cloudinary==1.34.0
dj-database-url==2.1.0
dj-rest-auth==2.1.9
Django==3.2.21
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
numpy==1.26.0
oauthlib==3.2.2
pandas==2.1.1
Pillow==10.0.0
plotly==5.17.0
psycopg2==2.9.7
PyJWT==2.8.0
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==1.3.1
sqlparse==0.4.4
tenacity==8.2.3
tzdata==2023.3
urllib3==1.26.15
